requirements.txt
pi-rc522==2.2.1
Pillow==10.0.0
RPi.GPIO==0.7.1
spidev==3.6
